ADD : Modify product
A product can now be modified trough the same form which is used for creating a product.
To modify a product, click on a product from the product list to show the pop-up. Then,
click on the modify button.

For further information, please consult : https://trello.com/c/czQVkl30/9-us-02-ajouter-et-supprimer-des-aliments-manuellement